+++ app/javascript/react/screens/App/common/twoPhaseConstants.js
